tox-requirements.txt
django-anymail==3.0
django-pytest==0.2.0
django-render-block==0.5
html2text==2016.5.29
mock==2.0.0
pytest==4.1.0
pytest-django==3.4.5
pytest-pythonpath==0.7.3
six==1.10.0
tox==2.3.1
django-ckeditor
